requirements.txt
discord.py==2.2.3
fastapi==0.95
python-dotenv==1.0.0
uvicorn==0.22.0
pydantic~=1.10.7
aiohttp~=3.8.4
loguru==0.7.0
aiofiles==23.1.0
python-multipart==0.0.6
pymysql==1.1.1